Use required-features to reduce the cfg mess. Also, rename the example to `cli` and use a subcommand since I already have other ideas for things to add.
